fix(minimax): isolate remains API fallback
